+++ src/types/signable-withdrawal.ts
@@ -5,46 +5,57 @@ interface SignableWithdrawalERC20 {
   type: TokenType.ERC20;
   data: {
     token_address: string;
-    decimals: number,
+    decimals: number;
   };
 }
 
 interface SignableWithdrawalERC721 {
   type: TokenType.ERC721;
   data: {
-    token_id: string,
+    token_id: string;
     token_address: string;
   };
 }
 
 type SignableWithdrawalEth = {
-  type: TokenType.ETH,
+  type: TokenType.ETH;
   data: {
-    decimals: number,
-  }
-}
+    decimals: number;
+  };
+};
 
-type SignableWithdrawalToken = SignableWithdrawalEth | SignableWithdrawalERC721 | SignableWithdrawalERC20
+type SignableWithdrawalToken =
+  | SignableWithdrawalEth
+  | SignableWithdrawalERC721
+  | SignableWithdrawalERC20;
 
 //SignableWithdrawal endpoint requires fields in snake_case format
-export function convertToSignableRequestFormat(token: TokenPrepareWithdrawal): SignableWithdrawalToken {
+export function convertToSignableRequestFormat(
+  token: TokenPrepareWithdrawal,
+): SignableWithdrawalToken {
   if (token.type === TokenType.ERC721) {
     return {
       type: TokenType.ERC721,
       data: {
-        token_id: token.data.tokenId,
-        token_address: token.data.tokenAddress,
+        token_id: token.tokenId,
+        token_address: token.tokenAddress,
       },
-    }
-  }
-  if (token.type === TokenType.ERC20) {
+    };
+  } else if (token.type === TokenType.ERC20) {
     return {
       type: TokenType.ERC20,
       data: {
-        decimals: token.data.decimals,
-        token_address: token.data.tokenAddress,
+        decimals: token.decimals,
+        token_address: token.tokenAddress,
       },
-    }
+    };
+  } else if (token.type === TokenType.ETH) {
+    return {
+      type: TokenType.ETH,
+      data: {
+        decimals: token.decimals,
+      },
+    };
   }
   return token;
 }