+++ src/types/withdrawal.ts
@@ -1,6 +1,4 @@
 import { TokenType } from './token';
-import { WalletConnection } from './index';
-import { WithdrawalsApi } from '../api';
 
 export interface ETHWithdrawal {
   type: TokenType.ETH;
@@ -8,47 +6,42 @@ export interface ETHWithdrawal {
 
 interface ETHPrepareWithdrawal {
   type: TokenType.ETH;
-  data: {
-    decimals: number
-  }
+  decimals: number;
 }
 
 export const ETH_PREPARE_WITHDRAWAL_DATA: ETHPrepareWithdrawal = {
   type: TokenType.ETH,
-  data: {
-    decimals: 18,
-  },
-}
+  decimals: 18,
+};
 
 export interface ERC721Withdrawal {
   type: TokenType.ERC721;
-  data: {
-    tokenId: string,
-    tokenAddress: string;
-  };
+  tokenId: string;
+  tokenAddress: string;
 }
 
 export interface ERC20Withdrawal {
   type: TokenType.ERC20;
-  data: {
-    tokenId: string,
-    tokenAddress: string;
-  };
+  tokenId: string;
+  tokenAddress: string;
 }
 
 export interface ERC20PrepareWithdrawal {
   type: TokenType.ERC20;
-  data: {
-    tokenAddress: string,
-    decimals: number,
-  };
+  tokenAddress: string;
+  decimals: number;
 }
 
+export type TokenWithdrawal =
+  | ETHWithdrawal
+  | ERC20Withdrawal
+  | ERC721Withdrawal;
+export type TokenPrepareWithdrawal =
+  | ETHPrepareWithdrawal
+  | ERC20PrepareWithdrawal
+  | ERC721Withdrawal;
 
-export type TokenWithdrawal = ETHWithdrawal | ERC20Withdrawal | ERC721Withdrawal;
-export type TokenPrepareWithdrawal = ETHPrepareWithdrawal | ERC20PrepareWithdrawal | ERC721Withdrawal;
-
-export type PrepareWithdrawalRequest =  {
+export type PrepareWithdrawalRequest = {
   token: TokenPrepareWithdrawal;
   quantity: string;
-}
+};